Updated to support original and cli workflows
