+++ hyperion-to-envi-converter/hyperion_data.py
@@ -1,0 +1,241 @@
+# ==================================================================================
+#                           CONSTANTS FILE
+#
+# DESCRIPTION: This file contains the needed data for converting Hyperion EO-1
+#              GeoTIFF files to ENVI.
+#
+# SOURCE:      https://developers.google.com/earth-engine/datasets/catalog/EO1_HYPERION
+#
+# ==================================================================================
+
+from enum import Enum
+from pydantic import BaseModel
+from ENVI import ByteOrderEnum
+
+BOM_MAP = {
+    b"II": ByteOrderEnum.LSF,
+    b"MM": ByteOrderEnum.MSF,
+}
+VNIR_SCALING_FACTOR = 40
+SWIR_SCALING_FACTOR = 80
+
+
+class BandRangeEnum(str, Enum):
+    VNIR = "VNIR"
+    SWIR = "SWIR"
+
+
+SCALING_MAP = {
+    BandRangeEnum.VNIR: VNIR_SCALING_FACTOR,
+    BandRangeEnum.SWIR: SWIR_SCALING_FACTOR,
+}
+
+
+class Band(BaseModel):
+    center_wavelength: float
+    fwhm: float
+    range: BandRangeEnum
+
+
+_bands_dict = {
+    "B008": {"center_wavelength": 426.8200, "fwhm": 11.3871, "range": "VNIR"},
+    "B009": {"center_wavelength": 436.9900, "fwhm": 11.3871, "range": "VNIR"},
+    "B010": {"center_wavelength": 447.1700, "fwhm": 11.3871, "range": "VNIR"},
+    "B011": {"center_wavelength": 457.3400, "fwhm": 11.3871, "range": "VNIR"},
+    "B012": {"center_wavelength": 467.5200, "fwhm": 11.3871, "range": "VNIR"},
+    "B013": {"center_wavelength": 477.6900, "fwhm": 11.3871, "range": "VNIR"},
+    "B014": {"center_wavelength": 487.8700, "fwhm": 11.3784, "range": "VNIR"},
+    "B015": {"center_wavelength": 498.0400, "fwhm": 11.3538, "range": "VNIR"},
+    "B016": {"center_wavelength": 508.2200, "fwhm": 11.3133, "range": "VNIR"},
+    "B017": {"center_wavelength": 518.3900, "fwhm": 11.2580, "range": "VNIR"},
+    "B018": {"center_wavelength": 528.5700, "fwhm": 11.1907, "range": "VNIR"},
+    "B019": {"center_wavelength": 538.7400, "fwhm": 11.1119, "range": "VNIR"},
+    "B020": {"center_wavelength": 548.9200, "fwhm": 11.0245, "range": "VNIR"},
+    "B021": {"center_wavelength": 559.0900, "fwhm": 10.9321, "range": "VNIR"},
+    "B022": {"center_wavelength": 569.2700, "fwhm": 10.8368, "range": "VNIR"},
+    "B023": {"center_wavelength": 579.4500, "fwhm": 10.7407, "range": "VNIR"},
+    "B024": {"center_wavelength": 589.6200, "fwhm": 10.6482, "range": "VNIR"},
+    "B025": {"center_wavelength": 599.8000, "fwhm": 10.5607, "range": "VNIR"},
+    "B026": {"center_wavelength": 609.9700, "fwhm": 10.4823, "range": "VNIR"},
+    "B027": {"center_wavelength": 620.1500, "fwhm": 10.4147, "range": "VNIR"},
+    "B028": {"center_wavelength": 630.3200, "fwhm": 10.3595, "range": "VNIR"},
+    "B029": {"center_wavelength": 640.5000, "fwhm": 10.3188, "range": "VNIR"},
+    "B030": {"center_wavelength": 650.6700, "fwhm": 10.2942, "range": "VNIR"},
+    "B031": {"center_wavelength": 660.8500, "fwhm": 10.2856, "range": "VNIR"},
+    "B032": {"center_wavelength": 671.0200, "fwhm": 10.2980, "range": "VNIR"},
+    "B033": {"center_wavelength": 681.2000, "fwhm": 10.3349, "range": "VNIR"},
+    "B034": {"center_wavelength": 691.3700, "fwhm": 10.3909, "range": "VNIR"},
+    "B035": {"center_wavelength": 701.5500, "fwhm": 10.4592, "range": "VNIR"},
+    "B036": {"center_wavelength": 711.7200, "fwhm": 10.5322, "range": "VNIR"},
+    "B037": {"center_wavelength": 721.9000, "fwhm": 10.6004, "range": "VNIR"},
+    "B038": {"center_wavelength": 732.0700, "fwhm": 10.6562, "range": "VNIR"},
+    "B039": {"center_wavelength": 742.2500, "fwhm": 10.6933, "range": "VNIR"},
+    "B040": {"center_wavelength": 752.4300, "fwhm": 10.7058, "range": "VNIR"},
+    "B041": {"center_wavelength": 762.6000, "fwhm": 10.7276, "range": "VNIR"},
+    "B042": {"center_wavelength": 772.7800, "fwhm": 10.7907, "range": "VNIR"},
+    "B043": {"center_wavelength": 782.9500, "fwhm": 10.8833, "range": "VNIR"},
+    "B044": {"center_wavelength": 793.1300, "fwhm": 10.9938, "range": "VNIR"},
+    "B045": {"center_wavelength": 803.3000, "fwhm": 11.1044, "range": "VNIR"},
+    "B046": {"center_wavelength": 813.4800, "fwhm": 11.1980, "range": "VNIR"},
+    "B047": {"center_wavelength": 823.6500, "fwhm": 11.2600, "range": "VNIR"},
+    "B048": {"center_wavelength": 833.8300, "fwhm": 11.2824, "range": "VNIR"},
+    "B049": {"center_wavelength": 844.0000, "fwhm": 11.2822, "range": "VNIR"},
+    "B050": {"center_wavelength": 854.1800, "fwhm": 11.2816, "range": "VNIR"},
+    "B051": {"center_wavelength": 864.3500, "fwhm": 11.2809, "range": "VNIR"},
+    "B052": {"center_wavelength": 874.5300, "fwhm": 11.2797, "range": "VNIR"},
+    "B053": {"center_wavelength": 884.7000, "fwhm": 11.2782, "range": "VNIR"},
+    "B054": {"center_wavelength": 894.8800, "fwhm": 11.2771, "range": "VNIR"},
+    "B055": {"center_wavelength": 905.0500, "fwhm": 11.2765, "range": "VNIR"},
+    "B056": {"center_wavelength": 915.2300, "fwhm": 11.2756, "range": "VNIR"},
+    "B057": {"center_wavelength": 925.4100, "fwhm": 11.2754, "range": "VNIR"},
+    "B077": {"center_wavelength": 912.4500, "fwhm": 11.0457, "range": "SWIR"},
+    "B078": {"center_wavelength": 922.5400, "fwhm": 11.0457, "range": "SWIR"},
+    "B079": {"center_wavelength": 932.6400, "fwhm": 11.0457, "range": "SWIR"},
+    "B080": {"center_wavelength": 942.7300, "fwhm": 11.0457, "range": "SWIR"},
+    "B081": {"center_wavelength": 952.8200, "fwhm": 11.0457, "range": "SWIR"},
+    "B082": {"center_wavelength": 962.9100, "fwhm": 11.0457, "range": "SWIR"},
+    "B083": {"center_wavelength": 972.9900, "fwhm": 11.0457, "range": "SWIR"},
+    "B084": {"center_wavelength": 983.0800, "fwhm": 11.0457, "range": "SWIR"},
+    "B085": {"center_wavelength": 993.1700, "fwhm": 11.0457, "range": "SWIR"},
+    "B086": {"center_wavelength": 1003.3000, "fwhm": 11.0457, "range": "SWIR"},
+    "B087": {"center_wavelength": 1013.3000, "fwhm": 11.0457, "range": "SWIR"},
+    "B088": {"center_wavelength": 1023.4000, "fwhm": 11.0451, "range": "SWIR"},
+    "B089": {"center_wavelength": 1033.4900, "fwhm": 11.0423, "range": "SWIR"},
+    "B090": {"center_wavelength": 1043.5900, "fwhm": 11.0372, "range": "SWIR"},
+    "B091": {"center_wavelength": 1053.6900, "fwhm": 11.0302, "range": "SWIR"},
+    "B092": {"center_wavelength": 1063.7900, "fwhm": 11.0218, "range": "SWIR"},
+    "B093": {"center_wavelength": 1073.8900, "fwhm": 11.0122, "range": "SWIR"},
+    "B094": {"center_wavelength": 1083.9900, "fwhm": 11.0013, "range": "SWIR"},
+    "B095": {"center_wavelength": 1094.0900, "fwhm": 10.9871, "range": "SWIR"},
+    "B096": {"center_wavelength": 1104.1900, "fwhm": 10.9732, "range": "SWIR"},
+    "B097": {"center_wavelength": 1114.1900, "fwhm": 10.9572, "range": "SWIR"},
+    "B098": {"center_wavelength": 1124.2800, "fwhm": 10.9418, "range": "SWIR"},
+    "B099": {"center_wavelength": 1134.3800, "fwhm": 10.9248, "range": "SWIR"},
+    "B100": {"center_wavelength": 1144.4800, "fwhm": 10.9065, "range": "SWIR"},
+    "B101": {"center_wavelength": 1154.5800, "fwhm": 10.8884, "range": "SWIR"},
+    "B102": {"center_wavelength": 1164.6800, "fwhm": 10.8696, "range": "SWIR"},
+    "B103": {"center_wavelength": 1174.7700, "fwhm": 10.8513, "range": "SWIR"},
+    "B104": {"center_wavelength": 1184.8700, "fwhm": 10.8335, "range": "SWIR"},
+    "B105": {"center_wavelength": 1194.9700, "fwhm": 10.8154, "range": "SWIR"},
+    "B106": {"center_wavelength": 1205.0700, "fwhm": 10.7979, "range": "SWIR"},
+    "B107": {"center_wavelength": 1215.1700, "fwhm": 10.7822, "range": "SWIR"},
+    "B108": {"center_wavelength": 1225.1700, "fwhm": 10.7663, "range": "SWIR"},
+    "B109": {"center_wavelength": 1235.2700, "fwhm": 10.7520, "range": "SWIR"},
+    "B110": {"center_wavelength": 1245.3600, "fwhm": 10.7385, "range": "SWIR"},
+    "B111": {"center_wavelength": 1255.4600, "fwhm": 10.7270, "range": "SWIR"},
+    "B112": {"center_wavelength": 1265.5600, "fwhm": 10.7174, "range": "SWIR"},
+    "B113": {"center_wavelength": 1275.6600, "fwhm": 10.7091, "range": "SWIR"},
+    "B114": {"center_wavelength": 1285.7600, "fwhm": 10.7022, "range": "SWIR"},
+    "B115": {"center_wavelength": 1295.8600, "fwhm": 10.6970, "range": "SWIR"},
+    "B116": {"center_wavelength": 1305.9600, "fwhm": 10.6946, "range": "SWIR"},
+    "B117": {"center_wavelength": 1316.0500, "fwhm": 10.6937, "range": "SWIR"},
+    "B118": {"center_wavelength": 1326.0500, "fwhm": 10.6949, "range": "SWIR"},
+    "B119": {"center_wavelength": 1336.1500, "fwhm": 10.6996, "range": "SWIR"},
+    "B120": {"center_wavelength": 1346.2500, "fwhm": 10.7058, "range": "SWIR"},
+    "B121": {"center_wavelength": 1356.3500, "fwhm": 10.7163, "range": "SWIR"},
+    "B122": {"center_wavelength": 1366.4500, "fwhm": 10.7283, "range": "SWIR"},
+    "B123": {"center_wavelength": 1376.5500, "fwhm": 10.7437, "range": "SWIR"},
+    "B124": {"center_wavelength": 1386.6500, "fwhm": 10.7612, "range": "SWIR"},
+    "B125": {"center_wavelength": 1396.7400, "fwhm": 10.7807, "range": "SWIR"},
+    "B126": {"center_wavelength": 1406.8400, "fwhm": 10.8034, "range": "SWIR"},
+    "B127": {"center_wavelength": 1416.9400, "fwhm": 10.8267, "range": "SWIR"},
+    "B128": {"center_wavelength": 1426.9400, "fwhm": 10.8534, "range": "SWIR"},
+    "B129": {"center_wavelength": 1437.0400, "fwhm": 10.8818, "range": "SWIR"},
+    "B130": {"center_wavelength": 1447.1400, "fwhm": 10.9110, "range": "SWIR"},
+    "B131": {"center_wavelength": 1457.2300, "fwhm": 10.9422, "range": "SWIR"},
+    "B132": {"center_wavelength": 1467.3300, "fwhm": 10.9743, "range": "SWIR"},
+    "B133": {"center_wavelength": 1477.4300, "fwhm": 11.0074, "range": "SWIR"},
+    "B134": {"center_wavelength": 1487.5300, "fwhm": 11.0414, "range": "SWIR"},
+    "B135": {"center_wavelength": 1497.6300, "fwhm": 11.0759, "range": "SWIR"},
+    "B136": {"center_wavelength": 1507.7300, "fwhm": 11.1108, "range": "SWIR"},
+    "B137": {"center_wavelength": 1517.8300, "fwhm": 11.1461, "range": "SWIR"},
+    "B138": {"center_wavelength": 1527.9200, "fwhm": 11.1811, "range": "SWIR"},
+    "B139": {"center_wavelength": 1537.9200, "fwhm": 11.2156, "range": "SWIR"},
+    "B140": {"center_wavelength": 1548.0200, "fwhm": 11.2496, "range": "SWIR"},
+    "B141": {"center_wavelength": 1558.1200, "fwhm": 11.2826, "range": "SWIR"},
+    "B142": {"center_wavelength": 1568.2200, "fwhm": 11.3146, "range": "SWIR"},
+    "B143": {"center_wavelength": 1578.3200, "fwhm": 11.3460, "range": "SWIR"},
+    "B144": {"center_wavelength": 1588.4200, "fwhm": 11.3753, "range": "SWIR"},
+    "B145": {"center_wavelength": 1598.5100, "fwhm": 11.4037, "range": "SWIR"},
+    "B146": {"center_wavelength": 1608.6100, "fwhm": 11.4302, "range": "SWIR"},
+    "B147": {"center_wavelength": 1618.7100, "fwhm": 11.4538, "range": "SWIR"},
+    "B148": {"center_wavelength": 1628.8100, "fwhm": 11.4760, "range": "SWIR"},
+    "B149": {"center_wavelength": 1638.8100, "fwhm": 11.4958, "range": "SWIR"},
+    "B150": {"center_wavelength": 1648.9000, "fwhm": 11.5133, "range": "SWIR"},
+    "B151": {"center_wavelength": 1659.0000, "fwhm": 11.5286, "range": "SWIR"},
+    "B152": {"center_wavelength": 1669.1000, "fwhm": 11.5404, "range": "SWIR"},
+    "B153": {"center_wavelength": 1679.2000, "fwhm": 11.5505, "range": "SWIR"},
+    "B154": {"center_wavelength": 1689.3000, "fwhm": 11.5580, "range": "SWIR"},
+    "B155": {"center_wavelength": 1699.4000, "fwhm": 11.5621, "range": "SWIR"},
+    "B156": {"center_wavelength": 1709.5000, "fwhm": 11.5634, "range": "SWIR"},
+    "B157": {"center_wavelength": 1719.6000, "fwhm": 11.5617, "range": "SWIR"},
+    "B158": {"center_wavelength": 1729.7000, "fwhm": 11.5563, "range": "SWIR"},
+    "B159": {"center_wavelength": 1739.7000, "fwhm": 11.5477, "range": "SWIR"},
+    "B160": {"center_wavelength": 1749.7900, "fwhm": 11.5346, "range": "SWIR"},
+    "B161": {"center_wavelength": 1759.8900, "fwhm": 11.5193, "range": "SWIR"},
+    "B162": {"center_wavelength": 1769.9900, "fwhm": 11.5002, "range": "SWIR"},
+    "B163": {"center_wavelength": 1780.0900, "fwhm": 11.4789, "range": "SWIR"},
+    "B164": {"center_wavelength": 1790.1900, "fwhm": 11.4548, "range": "SWIR"},
+    "B165": {"center_wavelength": 1800.2900, "fwhm": 11.4279, "range": "SWIR"},
+    "B166": {"center_wavelength": 1810.3800, "fwhm": 11.3994, "range": "SWIR"},
+    "B167": {"center_wavelength": 1820.4800, "fwhm": 11.3688, "range": "SWIR"},
+    "B168": {"center_wavelength": 1830.5800, "fwhm": 11.3366, "range": "SWIR"},
+    "B169": {"center_wavelength": 1840.5800, "fwhm": 11.3036, "range": "SWIR"},
+    "B170": {"center_wavelength": 1850.6800, "fwhm": 11.2696, "range": "SWIR"},
+    "B171": {"center_wavelength": 1860.7800, "fwhm": 11.2363, "range": "SWIR"},
+    "B172": {"center_wavelength": 1870.8700, "fwhm": 11.2007, "range": "SWIR"},
+    "B173": {"center_wavelength": 1880.9800, "fwhm": 11.1666, "range": "SWIR"},
+    "B174": {"center_wavelength": 1891.0700, "fwhm": 11.1333, "range": "SWIR"},
+    "B175": {"center_wavelength": 1901.1700, "fwhm": 11.1018, "range": "SWIR"},
+    "B176": {"center_wavelength": 1911.2700, "fwhm": 11.0714, "range": "SWIR"},
+    "B177": {"center_wavelength": 1921.3700, "fwhm": 11.0424, "range": "SWIR"},
+    "B178": {"center_wavelength": 1931.4700, "fwhm": 11.0155, "range": "SWIR"},
+    "B179": {"center_wavelength": 1941.5700, "fwhm": 10.9912, "range": "SWIR"},
+    "B180": {"center_wavelength": 1951.5700, "fwhm": 10.9698, "range": "SWIR"},
+    "B181": {"center_wavelength": 1961.6600, "fwhm": 10.9508, "range": "SWIR"},
+    "B182": {"center_wavelength": 1971.7600, "fwhm": 10.9355, "range": "SWIR"},
+    "B183": {"center_wavelength": 1981.8600, "fwhm": 10.9230, "range": "SWIR"},
+    "B184": {"center_wavelength": 1991.9600, "fwhm": 10.9139, "range": "SWIR"},
+    "B185": {"center_wavelength": 2002.0600, "fwhm": 10.9083, "range": "SWIR"},
+    "B186": {"center_wavelength": 2012.1500, "fwhm": 10.9069, "range": "SWIR"},
+    "B187": {"center_wavelength": 2022.2500, "fwhm": 10.9057, "range": "SWIR"},
+    "B188": {"center_wavelength": 2032.3500, "fwhm": 10.9013, "range": "SWIR"},
+    "B189": {"center_wavelength": 2042.4500, "fwhm": 10.8951, "range": "SWIR"},
+    "B190": {"center_wavelength": 2052.4500, "fwhm": 10.8854, "range": "SWIR"},
+    "B191": {"center_wavelength": 2062.5500, "fwhm": 10.8740, "range": "SWIR"},
+    "B192": {"center_wavelength": 2072.6500, "fwhm": 10.8591, "range": "SWIR"},
+    "B193": {"center_wavelength": 2082.7500, "fwhm": 10.8429, "range": "SWIR"},
+    "B194": {"center_wavelength": 2092.8400, "fwhm": 10.8242, "range": "SWIR"},
+    "B195": {"center_wavelength": 2102.9400, "fwhm": 10.8039, "range": "SWIR"},
+    "B196": {"center_wavelength": 2113.0400, "fwhm": 10.7820, "range": "SWIR"},
+    "B197": {"center_wavelength": 2123.1400, "fwhm": 10.7592, "range": "SWIR"},
+    "B198": {"center_wavelength": 2133.2400, "fwhm": 10.7342, "range": "SWIR"},
+    "B199": {"center_wavelength": 2143.3400, "fwhm": 10.7092, "range": "SWIR"},
+    "B200": {"center_wavelength": 2153.3400, "fwhm": 10.6834, "range": "SWIR"},
+    "B201": {"center_wavelength": 2163.4300, "fwhm": 10.6572, "range": "SWIR"},
+    "B202": {"center_wavelength": 2173.5300, "fwhm": 10.6312, "range": "SWIR"},
+    "B203": {"center_wavelength": 2183.6300, "fwhm": 10.6052, "range": "SWIR"},
+    "B204": {"center_wavelength": 2193.7300, "fwhm": 10.5803, "range": "SWIR"},
+    "B205": {"center_wavelength": 2203.8300, "fwhm": 10.5560, "range": "SWIR"},
+    "B206": {"center_wavelength": 2213.9300, "fwhm": 10.5328, "range": "SWIR"},
+    "B207": {"center_wavelength": 2224.0300, "fwhm": 10.5101, "range": "SWIR"},
+    "B208": {"center_wavelength": 2234.1200, "fwhm": 10.4904, "range": "SWIR"},
+    "B209": {"center_wavelength": 2244.2200, "fwhm": 10.4722, "range": "SWIR"},
+    "B210": {"center_wavelength": 2254.2200, "fwhm": 10.4552, "range": "SWIR"},
+    "B211": {"center_wavelength": 2264.3200, "fwhm": 10.4408, "range": "SWIR"},
+    "B212": {"center_wavelength": 2274.4200, "fwhm": 10.4285, "range": "SWIR"},
+    "B213": {"center_wavelength": 2284.5200, "fwhm": 10.4197, "range": "SWIR"},
+    "B214": {"center_wavelength": 2294.6100, "fwhm": 10.4129, "range": "SWIR"},
+    "B215": {"center_wavelength": 2304.7100, "fwhm": 10.4088, "range": "SWIR"},
+    "B216": {"center_wavelength": 2314.8100, "fwhm": 10.4077, "range": "SWIR"},
+    "B217": {"center_wavelength": 2324.9100, "fwhm": 10.4077, "range": "SWIR"},
+    "B218": {"center_wavelength": 2335.0100, "fwhm": 10.4077, "range": "SWIR"},
+    "B219": {"center_wavelength": 2345.1100, "fwhm": 10.4077, "range": "SWIR"},
+    "B220": {"center_wavelength": 2355.2100, "fwhm": 10.4077, "range": "SWIR"},
+    "B221": {"center_wavelength": 2365.2000, "fwhm": 10.4077, "range": "SWIR"},
+    "B222": {"center_wavelength": 2375.3000, "fwhm": 10.4077, "range": "SWIR"},
+    "B223": {"center_wavelength": 2385.4000, "fwhm": 10.4077, "range": "SWIR"},
+    "B224": {"center_wavelength": 2395.5000, "fwhm": 10.4077, "range": "SWIR"},
+}
+
+BANDS = {k: Band(**v) for k, v in _bands_dict.items()}
